Responsive UI Navbar/Homepage
updated the navbar and homepage to satisfy mobile first design principles. fixed alignment issues on smaller screens. incorporated a hamburger menu when viewing on a smaller screen and display all navbar components in the drawer component. adjusted components on the home page such as top carousel and testimonial carousel to allow changes depending on screen size.